Refactor ActivityTicker
