+++ app/javascript/components/track/activity-ticker/ExerciseWidget.tsx
@@ -1,0 +1,19 @@
+import React from 'react'
+export function ExerciseWidget({
+  exercise,
+}: {
+  exercise: Record<'exerciseUrl' | 'iconUrl' | 'title', string>
+}) {
+  return (
+    <span className="inline-flex">
+      &nbsp;to&nbsp;
+      <a
+        href={exercise.exerciseUrl}
+        className="flex flex-row items-center font-semibold text-prominentLinkColor"
+      >
+        <img src={exercise.iconUrl} className="w-[24px] h-[24px] mr-8" />
+        {exercise.title}
+      </a>
+    </span>
+  )
+}
