+++ app/javascript/components/track/activity-ticker/MetricText.ts
@@ -1,0 +1,5 @@
+export const METRIC_TEXT = {
+  publish_solution_metric: `published a new solution`,
+  open_pull_request_metric: `submitted a Pull Request on Exercism's GitHub`,
+  merge_pull_request_metric: `had a Pull Request merged on Exercism's GitHub`,
+}
