+++ app/javascript/components/track/activity-ticker/index.ts
@@ -1,0 +1,10 @@
+export { ExerciseWidget } from './ExerciseWidget'
+export { Handle } from './Handle'
+export { UserAvatar } from './UserAvatar'
+export { useActivityTicker } from './useActivityTicker'
+export {
+  ActivityTickerProps,
+  METRIC_TYPES,
+  allowedMetricTypes,
+} from './ActivityTicker.types'
+export { METRIC_TEXT } from './MetricText'
